+++ packages/globals/src/tokens/colors.doc.ts
