+++ packages/globals/src/tokens/letterSpacing.ts
