+++ packages/globals/src/tokens/maxSize.ts
