+++ packages/react/src/theme.ts
@@ -1,0 +1,1 @@
+export { theme } from "@optiaxiom/globals";
