--- packages/react/src/theme/index.ts
@@ -1,1 +1,0 @@
-export * from "./theme.css";
