+++ packages/react/src/tokens.ts
@@ -1,0 +1,1 @@
+export { tokens, tokensDark } from "@optiaxiom/globals";
